+++ CopilotKit/packages/runtime/src/lib/runtime/agui-action.ts
@@ -16,6 +16,7 @@ import { AbstractAgent } from "@ag-ui/client";
 import { CopilotKitError, CopilotKitErrorCode, parseJson } from "@copilotkit/shared";
 import { MetaEventInput } from "../../graphql/inputs/meta-event.input";
 import { GraphQLContext } from "../integrations/shared";
+import { CopilotContextInput } from "../../graphql/inputs/copilot-context.input";
 
 export function constructAGUIRemoteAction({
   logger,
@@ -25,6 +26,7 @@ export function constructAGUIRemoteAction({
   metaEvents,
   threadMetadata,
   nodeName,
+  context,
   graphqlContext,
 }: {
   logger: Logger;
@@ -34,6 +36,7 @@ export function constructAGUIRemoteAction({
   metaEvents?: MetaEventInput[];
   threadMetadata?: Record<string, any>;
   nodeName?: string;
+  context?: CopilotContextInput[];
   graphqlContext: GraphQLContext;
 }) {
   const action = {
@@ -45,6 +48,13 @@ export function constructAGUIRemoteAction({
       actionInputsWithoutAgents,
       threadId,
     }: RemoteAgentHandlerParams): Promise<Observable<RuntimeEvent>> => {
+      graphqlContext.request.signal.addEventListener(
+        "abort",
+        () => {
+          agent.abortRun();
+        },
+        { once: true }, // optional: fire only once
+      );
       logger.debug({ actionName: agent.agentId }, "Executing remote agent");
 
       const agentWireMessages = convertMessagesToAGUIMessage(messages);
@@ -92,6 +102,7 @@ export function constructAGUIRemoteAction({
         agent.legacy_to_be_removed_runAgentBridged({
           tools,
           forwardedProps,
+          context,
         }) as Observable<RuntimeEvent>
       ).pipe(
         mergeMap((event) => {