+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,47 @@
 # ui
 
+## 1.10.5-next.5
+
+### Patch Changes
+
+- Updated dependencies [6112a67]
+  - @copilotkit/react-core@1.10.5-next.5
+  - @copilotkit/runtime-client-gql@1.10.5-next.5
+  - @copilotkit/shared@1.10.5-next.5
+
+## 1.10.5-next.4
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.10.5-next.4
+- @copilotkit/react-core@1.10.5-next.4
+- @copilotkit/shared@1.10.5-next.4
+
+## 1.10.5-next.3
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.10.5-next.3
+- @copilotkit/react-core@1.10.5-next.3
+- @copilotkit/shared@1.10.5-next.3
+
+## 1.10.5-next.2
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.10.5-next.2
+- @copilotkit/react-core@1.10.5-next.2
+- @copilotkit/shared@1.10.5-next.2
+
+## 1.10.5-next.1
+
+### Patch Changes
+
+- Updated dependencies [b7bc3a0]
+  - @copilotkit/react-core@1.10.5-next.1
+  - @copilotkit/runtime-client-gql@1.10.5-next.1
+  - @copilotkit/shared@1.10.5-next.1
+
 ## 1.10.5-next.0
 
 ### Patch Changes